Save OpenNebula and Kubernetes configurations to DB

+++ src/main/kotlin/ru/ifmo/kirmanak/manager/models/exceptions/NoSuchConfigurationException.kt
@@ -4,4 +4,4 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(code = HttpStatus.BAD_REQUEST)
-class NoSuchConfigurationException(id: Int) : Exception("Configuration with id = $id was not found")
+class NoSuchConfigurationException(id: Long) : Exception("Configuration with id = $id was not found")
